Extract component for managing conenction lifecycle
